Kyrylo add mongodb to project

+++ src/main/java/kyrylo/delivery/com/deliveryproductmicroservice/services/CategoryService.java
@@ -1,12 +1,12 @@
 package kyrylo.delivery.com.deliveryproductmicroservice.services;
 
 import kyrylo.delivery.com.deliveryproductmicroservice.entities.Category;
+import kyrylo.delivery.com.deliveryproductmicroservice.exceptions.categoryExceptions.CategoryNotFoundException;
 import kyrylo.delivery.com.deliveryproductmicroservice.repositories.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CategoryService {
@@ -26,18 +26,21 @@ public class CategoryService {
         return categoryRepository.findAll();
     }
 
-    public Optional<Category> getCategoryById(Long id) {
-        return categoryRepository.findById(id);
+    public Category getCategoryById(String id) {
+        return categoryRepository.findById(id).orElseThrow(() ->
+                new CategoryNotFoundException("Category not found with id: " + id));
     }
 
-    public Optional<Category> updateCategory(Long id, Category categoryDetails) {
-        return getCategoryById(id).map(category -> {
-            category.setName(categoryDetails.getName());
-            return categoryRepository.save(category);
-        });
+    public Category updateCategory(String id, Category categoryDetails) {
+        Category category = getCategoryById(id);
+        category.setName(categoryDetails.getName());
+        return categoryRepository.save(category);
     }
 
-    public void deleteCategory(Long id) {
+    public void deleteCategory(String id) {
+        if(!categoryRepository.existsById(id))
+            throw new CategoryNotFoundException("Category not found with id: " + id);
+
         categoryRepository.deleteById(id);
     }
